UPSTREAM: ARM: s3c64xx: allow building without board support
Most of the code for the s3c64xx platform is only used when booting
with ATAGS based board files, but not when using device-tree.

This tries to identify all the s3c64xx specific code that is
unneeded when CONFIG_ATAGS is not set, so we can build a smaller
DT-only kernel if configured that way.

All board support is still left intact but now depends on the
CONFIG_ATAGS symbol that users may intentionally disable.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
cherry picked from commit b28210421c329766275b020ffeadce0a4e9d551f
Signed-off-by: Russ Weight <russell.h.weight@intel.com>