UPSTREAM: ARM: s3c64xx: use new adc/touchscreen driver
The old ADC and touchscreen drivers are not compatible with
multiplatform support, but we can use the exynos-adc driver
as a replacement.

This changes the common device creation functions for s3c64xx
(but not s3c24xx for now) to use the new driver. To do this,
we have to pass the interrupt resources in the opposite order
and pass the platform data in the adc device node.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
cherry picked from commit a829ae57f8b17bbebc7b9b2cbec99686b88a9e25
Signed-off-by: Russ Weight <russell.h.weight@intel.com>