UPSTREAM: ASoC: rockchip: i2s: change bclk and lrck according to sample rates
This patch sets the dividers autonomously.

when i2s works on master mode, and sample rates changed. We need to change
bclk and lrck at the same time for cpu internal side.

As the input source clock to the module is MCLK_I2S,
and by the divider of the module, the clock generator generates
SCLK and LRCK to transmitter and receiver.

BUG=chrome-os-partner:50756
TEST=Build and boot on rk3288

Change-Id: I8825a43ca9396b8127481d1e173f7d906994c709
Signed-off-by: Caesar Wang <wxt@rock-chips.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 2458c37779ddb91b4109949d86f5a5e193ba415b)