UPSTREAM: ASoC: arizona: In arizona_calc_fratio make new codecs the default case
This patch rearranges the switch statement in arizona_calc_fratio so
that older codecs are the special cases, with the default case
applying to newer codecs (WM8998 and later). This is preferable
because it avoids having to patch new cases in every time a new
codec is added.

Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
cherry picked from commit 3451eb485aee78f31a8dd127e3385f89946c813b
Signed-off-by: Russ Weight <russell.h.weight@intel.com>