UPSTREAM: ASoC: da7218: Correct BCLK inversion for DSP DAI format mode
By default the device latches data on the falling edge of the
BCLK in DSP mode, whereas the expectation for normal BCLK is to
latch on the rising edge. This updates the driver to invert the
BCLK configuration for DSP mode, to align with expected behaviour.

Signed-off-by: Adam Thomson <Adam.Thomson.Opensource@diasemi.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
cherry picked from commit f644eb62fe88a2414e5e5c1bd80e20c881f5d86a
Signed-off-by: Russ Weight <russell.h.weight@intel.com>