UPSTREAM: ASoC: davinci-mcasp: Discourage use of fck_parent for clock reparenting
The in-driver clock reparenting had been added when we did not had other
means to cleanly set the parent for the fck. Now we can use
assigned-clocks/assigned-clock-parents in DT binding. Print warning when
the fck_parent is present for McASP and recommend the switch to the proper
way to handle the clock selection.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
cherry picked from commit c670254f63fe09681b1c4c82d3d059c6cc04ce09
Signed-off-by: Russ Weight <russell.h.weight@intel.com>