UPSTREAM: ALSA: portman2x4 - use new parport device model
Modify portman driver to use the new parallel port device model.
The advantage of using the device model is that the device gets binded
to the hardware, we get the feature of hotplug, we can bind/unbind
the driver at runtime.
The changes are in the way the driver gets registered with the
parallel port subsystem and the temporary device to probe portman card
is removed and portman_probe() is used in the probe callback.

Signed-off-by: Sudip Mukherjee <sudip@vectorindia.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
cherry picked from commit e6a1b7e88046055989b8b59c1d648fda4e606f34
Signed-off-by: Russ Weight <russell.h.weight@intel.com>