UPSTREAM: ALSA: portman2x4: fix NULL pointer dereference
While registering pardev, the irq_func was also registered. As a result
when we tried to probe for the card, an interrupt was generated and in
the ISR we tried to dereference private_data. But private_data is still
NULL as we have not yet done portman_create(). Lets probe for the device
after card is created.

Fixes: e6a1b7e88046 ("ALSA: portman2x4 - use new parport device model")
Signed-off-by: Sudip Mukherjee <sudip.mukherjee@codethink.co.uk>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
cherry picked from commit 03367bf7fd30165bda3b037b3078cfa231ba33b7
Signed-off-by: Russ Weight <russell.h.weight@intel.com>