UPSTREAM: ASoC: rt5616: allow to build with CONFIG_SND_SOC_RT5616
This patch fixes that hasn't built the rt5616 driver with
'CONFIG_SND_SOC_RT5616=y' in .config.

The tristate is the prompt on the 'make menuconfig',
in other words, that can't show the prompt and select it
if we don't say what's the tristate.

Signed-off-by: Caesar Wang <wxt@rock-chips.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
cherry picked from commit 288bc356a881fcdce05f9782736a6a3d9db359a9
Signed-off-by: Russ Weight <russell.h.weight@intel.com>