UPSTREAM: ASoC: davinci-mcasp: dai format runtime reconfiguration
In case when the dai format is set via the dai_link the format
configuration happens once when the links are probed. If the McASP lose
context after this, the information will be lost and McASP will not going
to work correctly.
To overcome this issue, we save the fmt and set it within hw_params as
well.

Reported-by: Misael Lopez Cruz <misael.lopez@ti.com>
Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
cherry picked from commit 4a11ff260053b1097a5afa6c51e8d5e064aa88a7
Signed-off-by: Russ Weight <russell.h.weight@intel.com>