UPSTREAM: ALSA: hda - Fix forgotten HDMI monitor_present update
We forgot to copy monitor_present value when updating the ELD
information.  This won't change the ELD retrieval and the jack
notification behavior, but appears only in the proc output.   In that
sense, it's no fatal error, but a bug is a bug is a bug.

Cc: <stable@vger.kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
cherry picked from commit bd48128539ab89986b24ad08ecd3e027dd1993a1
Signed-off-by: Russ Weight <russell.h.weight@intel.com>